[development] refactor fileprocesslock, cleanups.
