+++ libs/file_processing.py
@@ -16,8 +16,9 @@ from config.constants import (ACCELEROMETER, ANDROID_LOG_FILE, API_TIME_FORMAT, 
     DATA_PROCESSING_NO_ERROR_STRING, IDENTIFIERS, IOS_LOG_FILE,
     SURVEY_DATA_FILES, SURVEY_TIMINGS, UPLOAD_FILE_TYPE_MAPPING, WIFI)
 from config.settings import CONCURRENT_NETWORK_OPS, FILE_PROCESS_PAGE_SIZE
-from database.data_access_models import ChunkRegistry, FileProcessLock, FileToProcess
+from database.data_access_models import ChunkRegistry, FileToProcess
 from database.survey_models import Survey
+from database.system_models import FileProcessLock
 from database.user_models import Participant
 from libs.s3 import s3_retrieve, s3_upload
